refactor(codegen): split algebra/expressions/types into codegen/ submodules
Phase 1 Batch 2 (Moves 4-6 from docs/PHASE1-PLAN.md).  Extract 28
handlers plus 7 helpers from latex_gen.py into three new mixin
modules.

Moves are byte-identical Fowler Move-Method / Extract-Class
refactorings.  Method bodies, signatures, names, and docstrings are
unchanged.  Mixin inheritance order is appended to LaTeXGenerator so
every handler remains reachable via the MRO.

* algebra.py — Restrict, Project, RelationRename, NaturalJoin,
  Divide, Group, GroupAggregate, Ungroup.  Includes the two private
  helpers _emit_attr_name and _generate_aggregator_clause that are
  called only from this family.
* expressions.py — Identifier, Number, UnaryOp, BinaryOp,
  Quantifier (logical / lambda / schema), Lambda, SetComprehension,
  FunctionApp, Range, Conditional, GuardedCases, GuardedBranch.
  Includes the quantifier / lambda chain helpers
  (_collect_lambda_chain, _generate_lambda_quantifier,
  _collect_quantifier_chain, _generate_logical_quantifier,
  _generate_schema_binding_latex, _generate_schema_quantifier),
  and the module-level _is_atomic_predicate function which is
  only used by _generate_unary_op.
* types.py — FunctionType, GenericInstantiation.

The CodegenDispatch TYPE_CHECKING block grows to declare the
cross-cutting helpers and class-vars still living on
LaTeXGenerator (paren-policy helpers, operator tables, indentation
helpers, the quantifier depth counter).  These declarations have
no runtime effect; they let mypy/pyright resolve self.X in moved
method bodies without modifying those bodies.

Gate 2 verified:
* make check — lint + markdown + format + mypy + pyright + 4700
  unit tests, all clean.
* make test-e2e — 159/159 examples/ .tex fixtures byte-identical.
* make refactor-diff — 190/190 examples/ + tests/bugs/ outputs
  byte-identical against main (the v1.3.1 baseline).

latex_gen.py shrinks from 4748 → 3400 lines.  Cumulative reduction
from the original 6864-line monolith: 50.5%.

Refs: docs/PHASE1-PLAN.md, docs/DESIGN-ROADMAP.md.
