+++ src/txt2tex/codegen/__init__.py
@@ -4,7 +4,10 @@ from __future__ import annotations
 
 from txt2tex.codegen import (
     _smoke,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    algebra,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    expressions,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    types,  # noqa: F401  # pyright: ignore[reportUnusedImport]
 )
